+++ src/utils/cleanDirectory.ts
@@ -1,0 +1,6 @@
+import fs from 'fs';
+
+export const cleanDirectory = (dir: string): void => {
+  fs.rmSync(dir, { force: true, recursive: true });
+  fs.mkdirSync(dir);
+};